requirements.txt
# --- Core Data Stack ---
# Using versions compatible with Python 3.12
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# --- The "Brain" (Modeling) ---
xgboost>=2.0.0
lifelines>=0.28.0
shap>=0.44.0

# --- The "Harvester" (Scraping & Cleaning) ---
playwright>=1.41.0
beautifulsoup4>=4.12.0
thefuzz>=0.22.0     # CRITICAL for matching Samadhaan names to Tofler
python-dotenv>=1.0.0

# --- File Handling ---
openpyxl>=3.1.0     # For .xlsx
xlrd>=2.0.1         # REQUIRED for your 'Report.xls' files

# --- The "Face" (API & Dashboard) ---
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.31.0

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0

# --- Development ---
pytest>=8.0.0
black>=24.0.0